[main] use http for local searxNG

+++ install.sh
@@ -5,9 +5,9 @@
 ######################################################################
 
 
-[ -f "$PWD"/dbw ] || echo "Error: No dbw file found"; exit 1
+[ ! -f "$PWD"/dbw ] && { echo "Error: No dbw file found"; exit 1 ; }
 
 # [ -d /usr/local/bin ] && sudo cp "$PWD"/dbw /usr/local/bin
 [ -d /usr/bin ]   && sudo cp "$PWD"/dbw.sh /usr/bin/dbw
 
-[ -f "${XDG_CONFIG_HOME:-$HOME/.config}/dbwdb.db"  ] ||  cp "$PWD"/example_database.db  "${XDG_CONFIG_HOME:-$HOME/.config}"
+[ ! -f "${XDG_CONFIG_HOME:-$HOME/.config}/dbwdb.db"  ] && cp "$PWD"/example_database.db  "${XDG_CONFIG_HOME:-$HOME/.config}"
